+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# unreleased
+
+* `WebOSMetaPlugin`: Fixed not inserting title into the output HTML.
+
 # 5.0.2 (September 16, 2022)
 
 * Pinned versions of dependencies as same as 5.0.0.